Changed a lot of things
Sort of working now! Still a few bugs to work out, but this script will take a fresh drive, create a partition, create a directory, assign appropriate permissions, and lets you install games to that partition.

A few problems:
1. It SHOULD just format your current drive to make a new partition of the size you indicate.
2. It SHOULD auto-mount the drive on reboot, but doesn't (this might be because of my testing rig)
3. It SHOULD create a mount point for your game directory, but it doesn't.

Still working on these things, but it at least sort of works.